Task 5/requirements.txt
streamlit>=1.33.0
requests>=2.31.0
pillow>=10.2.0
